+++ src/main/java/com/eventopia/review/ReviewDetailC.java
@@ -18,48 +18,10 @@ public class ReviewDetailC extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response)
 			throws ServletException, IOException {
 
-		 showPost(request, response);
-
+		ReviewDAO.RDAO.showPost(request, response);
+		request.setAttribute("content", "jsp/review/reviewDetail.jsp");
+		request.getRequestDispatcher("index.jsp").forward(request, response);
 	}
 
-	public void showPost(HttpServletRequest request, HttpServletResponse response) throws IOException {
-        int postId = Integer.parseInt(request.getParameter("id"));
-
-        String sql = "SELECT title, content FROM review_post WHERE id = ?";
-        try (Connection con = DBManager.connect();
-             PreparedStatement pstmt = con.prepareStatement(sql)) {
-            pstmt.setInt(1, postId);
-
-            ResultSet rs = pstmt.executeQuery();
-            if (rs.next()) {
-                String title = rs.getString("title");
-                String content = rs.getString("content");
-                String img = rs.getString("img");
-                String sub = rs.getString("sub");
-                String date = rs.getString("date");
-
-
-                
-response.setContentType("text/html; charset=UTF-8");
-                response.getWriter().write("<html>");
-                response.getWriter().write("<head>");
-                response.getWriter().write("<title>" + title + "</title>");
-                response.getWriter().write("<style>");
-                response.getWriter().write("body { display: flex; flex-direction: column; align-items: center; justify-content: center; height: 100vh; margin: 0; }");
-                response.getWriter().write("h1 { text-align: center; }");
-                response.getWriter().write("div { text-align: center; max-width: 600px; }");
-                response.getWriter().write("</style>");
-                response.getWriter().write("</head>");
-                response.getWriter().write("<body>");
-                response.getWriter().write("<h1>" + title + "</h1>");
-                response.getWriter().write("<div>" + content + "</div>");
-                response.getWriter().write("</body>");
-                response.getWriter().write("</html>");
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
 }
-
